+++ app/class_day/morning.py
@@ -1,0 +1,10 @@
+class Morning:
+    def __init__(self):
+        self._period = "morning"
+        self._dishes = ["eggs", "toast", "coffee"]
+        
+    def get_period(self):
+        return self._period
+    
+    def get_dishes(self):
+        return self._dishes
